refactor: move ProviderAdresses schemas & types to separate file
